Update BaseSDTrainProcess.py
Please also see my proposed changes to the optimizer.py

You can dynamically import user-defined optimizers in optimizer.py, but because the import is case sensitive you need to allow case to remain as inputted by the user.

It doesn't seem necessary to force the optimizer_type to lower here anyway because the optimizer.py is already doing this each time it checks the value of optimizer_type.  as far as I can tell this is redundant anyway.